requirements.txt
streamlit
langchain
langchain-ollama
langchain-community
python-dotenv
unstructured
unstructured[pdf]
onnxruntime
faiss-cpu
tiktoken
envyaml
langchain-huggingface